+++ packages/lambda/src/functions/helpers/render-has-audio-video.ts
@@ -1,4 +1,4 @@
-import {BrowserSafeApis} from '@remotion/renderer/client';
+import {NoReactAPIs} from '@remotion/renderer/pure';
 import type {CloudProvider} from '@remotion/serverless';
 import type {RenderMetadata} from '@remotion/serverless/client';
 
@@ -12,10 +12,10 @@ export const lambdaRenderHasAudioVideo = <Provider extends CloudProvider>(
 		throw new Error('Cannot merge stills');
 	}
 
-	const support = BrowserSafeApis.codecSupportsMedia(renderMetadata.codec);
+	const support = NoReactAPIs.codecSupportsMedia(renderMetadata.codec);
 
 	const hasVideo = renderMetadata
-		? !BrowserSafeApis.isAudioCodec(renderMetadata.codec)
+		? !NoReactAPIs.isAudioCodec(renderMetadata.codec)
 		: false;
 	const hasAudio = renderMetadata
 		? !renderMetadata.muted && support.audio
