+++ packages/lambda/src/functions/renderer.ts
@@ -5,6 +5,7 @@ import type {
 	OnArtifact,
 } from '@remotion/renderer';
 import {RenderInternals} from '@remotion/renderer';
+import {NoReactAPIs} from '@remotion/renderer/pure';
 import type {
 	CloudProvider,
 	OnStream,
@@ -175,7 +176,7 @@ const renderHandler = async <Provider extends CloudProvider>({
 	const audioOutputLocation =
 		willRenderAudioEval === 'no'
 			? null
-			: RenderInternals.isAudioCodec(params.codec)
+			: NoReactAPIs.isAudioCodec(params.codec)
 				? null
 				: audioExtension
 					? path.join(outdir, `${chunk}.${audioExtension}`)
@@ -358,7 +359,7 @@ const renderHandler = async <Provider extends CloudProvider>({
 	if (videoOutputLocation) {
 		const videoChunkTimer = timer('Sending main chunk', params.logLevel);
 		await onStream({
-			type: RenderInternals.isAudioCodec(params.codec)
+			type: NoReactAPIs.isAudioCodec(params.codec)
 				? 'audio-chunk-rendered'
 				: 'video-chunk-rendered',
 			payload: fs.readFileSync(videoOutputLocation),
